github-profile/automate.py
import aiohttp
from bs4 import BeautifulSoup

from custom_logger import CustomLogger


class Automate:
    def __init__(self) -> None:
        self.logger = CustomLogger()

    async def flash_news_data(self) -> list[dict[str, str]]:
        data: list[dict[str, str]] = []
        async with aiohttp.ClientSession() as session, session.get("https://news.yahoo.co.jp/flash") as res:
            [
                data.append({
                    "title": str(tag["aria-label"]),
                    "link": str(tag["href"]),
                    "image": str(tag.find("source", {"type": "image/webp"})["srcset"]),  # type: ignore  # noqa: PGH003
                })
                for tag in BeautifulSoup(await res.text(), "html.parser").select("#contentsWrap > div > div > a")
            ]
        self.logger.debug(data)
        return data
