github-profile/requirements.txt
beautifulsoup4==4.13.4
python-dotenv==1.1.1
aiohttp==3.12.13
Jinja2==3.1.6
